Reached 63 epochs, almost converging
Results look bad, but I will try to get the training loss to be less than the validation loss, it appears to be getting there. I may have to run past 100 epochs though, then we will start to see results.

This is mirroring my previous example of running the exact same model without the quantum aspect, it took a while for the validation loss to overtake the training loss. I will give it more time. Hopefully it will end up working and the trend will not reverse.