114 Epochs. ROC and PR AUC improved. Still errors
The distribution is still bad. I will run it until 150 to see if it fixes anything. At least the ROC and PR AUC curves are starting to improve now that the training and validation loss are where they should be. I will run it some more tonight and hopefully come to the office Saturday to debug it again