2 gates test, improvement
I wanted to test with 2 gates in this improved version, it seems to be far better than the previous example as it actually returns something that resembles a correct answer, hwoever it still has some issues with the training/validation curve. I will play around with more gates to hopefully improve the model, please note that it will take longer. The main purpose of this experiment was to test if this method would work and it seems to be a significant improvement from last time.