WIP: Will play around with Alpha/Gamma values more
So far GNN outperforms QGNN when implementing Focal Loss. I suspect this is because the Alpha and Gamma values are not optimized for the QGNN, as I had selected the Alpha and Gamma values specifically because they optimized QGNN results. This was a test run at first, but it shows extremely promising results. I will play around with the QGNN model to render faster results.